Simpler cache system for symbol-tree-info
