+++ src/Features/Core/Portable/AddImport/SearchScopes/MetadataSymbolsSearchScope.cs
@@ -49,7 +49,7 @@ namespace Microsoft.CodeAnalysis.AddImport
             protected override async Task<ImmutableArray<ISymbol>> FindDeclarationsAsync(
                 SymbolFilter filter, SearchQuery searchQuery)
             {
-                var service = _solution.Services.GetService<ISymbolTreeInfoCacheService>();
+                var service = _solution.Services.GetService<SymbolTreeInfoCacheService>();
                 var info = await service.TryGetMetadataSymbolTreeInfoAsync(_solution, _metadataReference, CancellationToken).ConfigureAwait(false);
                 if (info == null)
                 {
