+++ src/Features/Core/Portable/IncrementalCaches/SymbolTreeInfoIncrementalAnalyzer.cs
@@ -22,45 +22,22 @@ namespace Microsoft.CodeAnalysis.IncrementalCaches
     {
         private class SymbolTreeInfoIncrementalAnalyzer : IncrementalAnalyzerBase
         {
-            // Shared with SymbolTreeInfoCacheService.  We populate the values, they read from them.
+            private readonly SymbolTreeInfoCacheService _cacheService;
 
-            private readonly ConcurrentDictionary<ProjectId, SymbolTreeInfo> _projectIdToInfo;
-            private readonly ConcurrentDictionary<MetadataId, SymbolTreeInfoCacheService.MetadataInfo> _metadataIdToInfo;
-
-            public SymbolTreeInfoIncrementalAnalyzer(
-                ISymbolTreeInfoCacheService cacheService)
+            public SymbolTreeInfoIncrementalAnalyzer(SymbolTreeInfoCacheService cacheService)
             {
-                var concreteCacheService = (SymbolTreeInfoCacheService)cacheService;
-                _projectIdToInfo = concreteCacheService.ProjectIdToInfo;
-                _metadataIdToInfo = concreteCacheService.MetadataIdToInfo;
+                _cacheService = cacheService;
             }
 
             private static bool SupportAnalysis(Project project)
                 => project.SupportsCompilation;
 
-            public override async Task AnalyzeDocumentAsync(Document document, SyntaxNode bodyOpt, InvocationReasons reasons, CancellationToken cancellationToken)
+            public override Task AnalyzeDocumentAsync(Document document, SyntaxNode bodyOpt, InvocationReasons reasons, CancellationToken cancellationToken)
             {
                 if (!SupportAnalysis(document.Project))
-                    return;
+                    return Task.CompletedTask;
 
-                if (bodyOpt != null)
-                {
-                    // This was a method body edit.  We can reuse the existing SymbolTreeInfo if
-                    // we have one.  We can't just bail out here as the change in the document means
-                    // we'll have a new checksum.  We need to get that new checksum so that our
-                    // cached information is valid.
-                    if (_projectIdToInfo.TryGetValue(document.Project.Id, out var cachedInfo))
-                    {
-                        var checksum = await SymbolTreeInfo.GetSourceSymbolsChecksumAsync(
-                            document.Project, cancellationToken).ConfigureAwait(false);
-
-                        var newInfo = cachedInfo.WithChecksum(checksum);
-                        _projectIdToInfo[document.Project.Id] = newInfo;
-                        return;
-                    }
-                }
-
-                await UpdateSymbolTreeInfoAsync(document.Project, cancellationToken).ConfigureAwait(false);
+                return _cacheService.AnalyzeDocumentAsync(document, bodyOpt, cancellationToken);
             }
 
             public override Task AnalyzeProjectAsync(Project project, bool semanticsChanged, InvocationReasons reasons, CancellationToken cancellationToken)
@@ -68,130 +45,14 @@ namespace Microsoft.CodeAnalysis.IncrementalCaches
                 if (!SupportAnalysis(project))
                     return Task.CompletedTask;
 
-                return UpdateSymbolTreeInfoAsync(project, cancellationToken);
-            }
-
-            private async Task UpdateSymbolTreeInfoAsync(Project project, CancellationToken cancellationToken)
-            {
-                Debug.Assert(SupportAnalysis(project));
-
-                // Produce the indices for the source and metadata symbols in parallel.
-                using var _ = ArrayBuilder<Task>.GetInstance(out var tasks);
-
-                tasks.Add(Task.Run(() => this.UpdateSourceSymbolTreeInfoAsync(project, cancellationToken), cancellationToken));
-                tasks.Add(Task.Run(() => this.UpdateReferencesAsync(project, cancellationToken), cancellationToken));
-
-                await Task.WhenAll(tasks).ConfigureAwait(false);
-            }
-
-            private async Task UpdateSourceSymbolTreeInfoAsync(Project project, CancellationToken cancellationToken)
-            {
-                var checksum = await SymbolTreeInfo.GetSourceSymbolsChecksumAsync(project, cancellationToken).ConfigureAwait(false);
-                if (!_projectIdToInfo.TryGetValue(project.Id, out var projectInfo) ||
-                    projectInfo.Checksum != checksum)
-                {
-                    projectInfo = await SymbolTreeInfo.GetInfoForSourceAssemblyAsync(
-                        project, checksum, loadOnly: false, cancellationToken).ConfigureAwait(false);
-
-                    Contract.ThrowIfNull(projectInfo);
-                    Contract.ThrowIfTrue(projectInfo.Checksum != checksum, "If we computed a SymbolTreeInfo, then its checksum much match our checksum.");
-
-                    // Mark that we're up to date with this project.  Future calls with the same 
-                    // semantic version can bail out immediately.
-                    _projectIdToInfo[project.Id] = projectInfo;
-                }
-            }
-
-            private async Task UpdateReferencesAsync(Project project, CancellationToken cancellationToken)
-            {
-                // Process all metadata references. If it remote workspace, do this in parallel.
-                using var pendingTasks = new TemporaryArray<Task>();
-
-                foreach (var reference in project.MetadataReferences)
-                {
-                    if (reference is not PortableExecutableReference portableExecutableReference)
-                        continue;
-
-                    if (cancellationToken.IsCancellationRequested)
-                    {
-                        // Break out of this loop to make sure other pending operations process cancellation before
-                        // returning.
-                        break;
-                    }
-
-                    var updateTask = UpdateReferenceAsync(_metadataIdToInfo, project, portableExecutableReference, cancellationToken);
-                    if (updateTask.Status != TaskStatus.RanToCompletion)
-                        pendingTasks.Add(updateTask);
-                }
-
-                if (pendingTasks.Count > 0)
-                {
-                    // If any update operations did not complete synchronously (including any cancelled operations),
-                    // wait for them to complete now.
-                    await Task.WhenAll(pendingTasks.ToImmutableAndClear()).ConfigureAwait(false);
-                }
-
-                // ⚠ This local function must be 'async' to ensure exceptions are captured in the resulting task and
-                // not thrown directly to the caller.
-                static async Task UpdateReferenceAsync(
-                    ConcurrentDictionary<MetadataId, SymbolTreeInfoCacheService.MetadataInfo> metadataIdToInfo,
-                    Project project,
-                    PortableExecutableReference reference,
-                    CancellationToken cancellationToken)
-                {
-                    var metadataId = SymbolTreeInfo.GetMetadataIdNoThrow(reference);
-                    if (metadataId == null)
-                        return;
-
-                    // 🐉 PERF: GetMetadataChecksum indirectly uses a ConditionalWeakTable. This call is intentionally
-                    // placed before the first 'await' of this asynchronous method to ensure it executes in the
-                    // synchronous portion of the caller. https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1270250
-                    var checksum = SymbolTreeInfo.GetMetadataChecksum(project.Solution, reference, cancellationToken);
-                    if (!metadataIdToInfo.TryGetValue(metadataId, out var metadataInfo) ||
-                        metadataInfo.SymbolTreeInfo.Checksum != checksum)
-                    {
-                        var info = await SymbolTreeInfo.GetInfoForMetadataReferenceAsync(
-                            project.Solution, reference, checksum, loadOnly: false, cancellationToken: cancellationToken).ConfigureAwait(false);
-
-                        Contract.ThrowIfNull(info);
-                        Contract.ThrowIfTrue(info.Checksum != checksum, "If we computed a SymbolTreeInfo, then its checksum much match our checksum.");
-
-                        // Note, getting the info may fail (for example, bogus metadata).  That's ok.  
-                        // We still want to cache that result so that don't try to continuously produce
-                        // this info over and over again.
-                        metadataInfo = new SymbolTreeInfoCacheService.MetadataInfo(info, metadataInfo.ReferencingProjects ?? new HashSet<ProjectId>());
-                        metadataIdToInfo[metadataId] = metadataInfo;
-                    }
-
-                    // Keep track that this dll is referenced by this project.
-                    lock (metadataInfo.ReferencingProjects)
-                    {
-                        metadataInfo.ReferencingProjects.Add(project.Id);
-                    }
-                }
+                return _cacheService.UpdateSymbolTreeInfoAsync(project, cancellationToken);
             }
 
             public override Task RemoveProjectAsync(ProjectId projectId, CancellationToken cancellationToken)
             {
-                _projectIdToInfo.TryRemove(projectId, out _);
-                RemoveMetadataReferences(projectId);
+                _cacheService.RemoveProject(projectId);
 
                 return Task.CompletedTask;
-            }
-
-            private void RemoveMetadataReferences(ProjectId projectId)
-            {
-                foreach (var (id, info) in _metadataIdToInfo.ToArray())
-                {
-                    lock (info.ReferencingProjects)
-                    {
-                        info.ReferencingProjects.Remove(projectId);
-
-                        // If this metadata dll isn't referenced by any project.  We can just dump it.
-                        if (info.ReferencingProjects.Count == 0)
-                            _metadataIdToInfo.TryRemove(id, out _);
-                    }
-                }
             }
         }
     }
