+++ src/Workspaces/Core/Portable/FindSymbols/SymbolTree/SymbolTreeInfoCacheService.MetadataInfo.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-
 using System.Collections.Generic;
 using Roslyn.Utilities;
